renamed test utils folder
